feat: AbstractRule.class 골격구현클래스 구현
