+++ src/main/java/chess/domain/RuleImpl/King.java
@@ -2,12 +2,12 @@ package chess.domain.RuleImpl;
 
 import chess.domain.Position;
 
-public class King implements Rule {
+public class King extends AbstractRule {
     private static King INSTANCE = new King();
     private static final double SCORE = 0;
 
     private King() {
-
+        super(SCORE);
     }
 
     public static King getInstance() {
@@ -25,15 +25,5 @@ public class King implements Rule {
 
     private boolean isLengthTwo(final Position origin, final Position target) {
         return origin.isMoveAnyWhereSum(target, 2) && origin.isMoveAnyWhereSub(target, 0);
-    }
-
-    @Override
-    public boolean isValidAttack(final Position origin, final Position target) {
-        return isValidMove(origin, target);
-    }
-
-    @Override
-    public double getScore() {
-        return SCORE;
     }
 }
