+++ src/main/java/chess/domain/RuleImpl/Knight.java
@@ -2,12 +2,12 @@ package chess.domain.RuleImpl;
 
 import chess.domain.Position;
 
-public class Knight implements Rule {
+public class Knight extends AbstractRule {
     private static Knight INSTANCE = new Knight();
     private static final double SCORE = 2.5;
 
     private Knight() {
-
+        super(SCORE);
     }
 
     public static Knight getInstance() {
@@ -17,15 +17,5 @@ public class Knight implements Rule {
     @Override
     public boolean isValidMove(final Position origin, final Position target) {
         return origin.isMoveAnyWhereSum(target, 3) && origin.isMoveAnyWhereSub(target, 1);
-    }
-
-    @Override
-    public boolean isValidAttack(final Position origin, final Position target) {
-        return isValidMove(origin, target);
-    }
-
-    @Override
-    public double getScore() {
-        return SCORE;
     }
 }
